Struct renaming::ctx::Ctx

pub struct Ctx {
    bound: Vec<Vec<Ident>>,
}

Fields
bound: Vec<Vec<Ident>>
Implementations

impl Ctx

pub fn empty() -> Self

pub fn bound(&self, idx: Idx) -> Ident

pub fn bind<T, F>(&mut self, name: Ident, f: F) -> T where
F: Fn(&mut Ctx) -> T,

Bind a single name

pub fn bind_fold<T, I: Iterator<Item = T>, O1, O2, F1, F2>(
&mut self,
iter: I,
acc: O1,
f_acc: F1,
f_inner: F2
) -> O2 where
T: Named,
F1: Fn(&mut Ctx, O1, T) -> O1,
F2: FnOnce(&mut Ctx, O1) -> O2,

Bind an iterator iter of binders

Fold the iterator and consume the result under the inner context with all binders in scope.

iter - An iterator of binders implementing Named.

acc - Accumulator for folding the iterator

f_acc - Accumulator function run for each binder

f_inner - Inner function computing the final result under the context of all binders

fn level_inc_fst(&mut self)

Increment the first component of the current De-Bruijn level

fn level_dec_fst(&mut self)

Decrement the first component of the current De-Bruijn level

fn push(&mut self, name: Ident)

Push a binder contained in a binder list, incrementing the second dimension of the current De Bruijn level

fn pop(&mut self)

Push a binder contained in a binder list, decrementing the second dimension of the current De Bruijn level